Cleaned up Itinerary page file
