+++ frontend/src/pages/itinerary/ItineraryPage.js
@@ -1,143 +1,20 @@
-import React, { useState } from 'react';
-import Accordion from '@mui/material/Accordion';
-import AccordionSummary from '@mui/material/AccordionSummary';
-import AccordionDetails from '@mui/material/AccordionDetails';
-import Typography from '@mui/material/Typography';
-import Button from '@mui/material/Button';
-import Dialog from '@mui/material/Dialog';
-import DialogTitle from '@mui/material/DialogTitle';
-import DialogContent from '@mui/material/DialogContent';
-import DialogContentText from '@mui/material/DialogContentText';
-import DialogActions from '@mui/material/DialogActions';
-import TextField from '@mui/material/TextField';
-import DatePickerComponent from '../../components/ItineraryForm/DatePickerComponent'; // Import your DatePickerComponent
-
-
-const ActivityAccordion = ({ activity, onEdit, onRemove }) => {
-  const [isConfirmDialogOpen, setConfirmDialogOpen] = useState(false);
-
-  const handleRemoveClick = () => {
-    setConfirmDialogOpen(true);
-  };
-
-  const handleConfirmRemove = () => {
-    setConfirmDialogOpen(false);
-    onRemove();
-  };
-
-  const handleCancelRemove = () => {
-    setConfirmDialogOpen(false);
-  };
-
-  return (
-    <Accordion>
-      <AccordionSummary>
-        <Typography>{activity}</Typography>
-      </AccordionSummary>
-      <AccordionDetails>
-        <div>
-          <Button onClick={onEdit}>Edit</Button>
-          <Button onClick={handleRemoveClick}>Remove</Button>
-        </div>
-
-        {/* Remove Confirmation Dialog */}
-        <Dialog open={isConfirmDialogOpen} onClose={handleCancelRemove}>
-          <DialogTitle>Confirm Removal</DialogTitle>
-          <DialogContent>
-            <DialogContentText>Are you sure you want to remove this activity?</DialogContentText>
-          </DialogContent>
-          <DialogActions>
-            <Button onClick={handleCancelRemove} color="primary">
-              Cancel
-            </Button>
-            <Button onClick={handleConfirmRemove} color="primary">
-              Confirm
-            </Button>
-          </DialogActions>
-        </Dialog>
-      </AccordionDetails>
-    </Accordion>
-  );
-};
-
-const ItineraryAccordion = ({ day, activities, onAddActivity, onEditActivity, onRemoveActivity }) => {
-  const [isAddDialogOpen, setAddDialogOpen] = useState(false);
-  const [newActivity, setNewActivity] = useState('');
-
-  const handleAddClick = () => {
-    setAddDialogOpen(true);
-  };
-
-  const handleConfirmAdd = () => {
-    setAddDialogOpen(false);
-    if (newActivity) {
-      onAddActivity(day, newActivity);
-      setNewActivity('');
-    }
-  };
-
-  const handleCancelAdd = () => {
-    setAddDialogOpen(false);
-  };
-
-  return (
-    <Accordion>
-      <AccordionSummary>
-        <Typography>Day {day}</Typography>
-      </AccordionSummary>
-      <AccordionDetails>
-        <div>
-          <ul>
-            {activities.map((activity, index) => (
-              <li key={index}>
-                <ActivityAccordion
-                  activity={activity}
-                  onEdit={() => onEditActivity(day, index)}
-                  onRemove={() => onRemoveActivity(day, index)}
-                />
-              </li>
-            ))}
-          </ul>
-          <Button onClick={handleAddClick}>Add Activity</Button>
-
-          {/* Add Activity Dialog */}
-          <Dialog open={isAddDialogOpen} onClose={handleCancelAdd}>
-            <DialogTitle>Add New Activity</DialogTitle>
-            <DialogContent>
-              <DialogContentText>Enter the details for the new activity:</DialogContentText>
-              <TextField
-                autoFocus
-                margin="dense"
-                id="activity"
-                label="Activity"
-                type="text"
-                fullWidth
-                value={newActivity}
-                onChange={(e) => setNewActivity(e.target.value)}
-              />
-            </DialogContent>
-            <DialogActions>
-              <Button onClick={handleCancelAdd} color="primary">
-                Cancel
-              </Button>
-              <Button onClick={handleConfirmAdd} color="primary">
-                Add
-              </Button>
-            </DialogActions>
-          </Dialog>
-        </div>
-      </AccordionDetails>
-    </Accordion>
-  );
-};
+import React, { useState } from "react";
+import Accordion from "@mui/material/Accordion";
+import AccordionSummary from "@mui/material/AccordionSummary";
+import AccordionDetails from "@mui/material/AccordionDetails";
+import Typography from "@mui/material/Typography";
+import Button from "@mui/material/Button";
+import TextField from "@mui/material/TextField";
+import DatePickerComponent from "../../components/ItineraryForm/DatePickerComponent";
+import ItineraryAccordion from "../../components/ItineraryForm/ItineraryAccordion";
 
 const possibleActivities = [
-  'Explore the city',
-  'Visit a museum',
-  'Dinner at a local restaurant',
-  'City park exploration',
-  'Relax at the beach',
-  'Sunset Cruise',
+  "Explore the city",
+  "Visit a museum",
+  "Dinner at a local restaurant",
+  "City park exploration",
+  "Relax at the beach",
+  "Sunset Cruise",
   // Add more activities as needed
 ];
 
@@ -147,9 +24,9 @@ const getRandomActivity = () => {
 };
 
 const ItineraryPage = () => {
-  const [currentCity] = useState('Your City');
+  const [currentCity] = useState("Your City");
   const [itinerary, setItinerary] = useState([]);
-  const [currentStep, setCurrentStep] = useState('datePicker');
+  const [currentStep, setCurrentStep] = useState("datePicker");
 
   const [selectedStartDate, setSelectedStartDate] = useState(null);
   const [selectedEndDate, setSelectedEndDate] = useState(null);
@@ -163,14 +40,17 @@ const ItineraryPage = () => {
   };
 
   const handleDateSelectionComplete = () => {
-    const numberOfDays = calculateNumberOfDays(selectedStartDate, selectedEndDate);
+    const numberOfDays = calculateNumberOfDays(
+      selectedStartDate,
+      selectedEndDate
+    );
 
     if (numberOfDays > 0) {
       generateRandomItinerary(numberOfDays);
-      setCurrentStep('itinerary');
+      setCurrentStep("itinerary");
     } else {
       // Handle invalid date range
-      alert('Please select a valid date range.');
+      alert("Please select a valid date range.");
     }
   };
 
@@ -180,15 +60,19 @@ const ItineraryPage = () => {
   };
 
   const generateRandomItinerary = (numberOfDays) => {
-    setItinerary(Array.from({ length: numberOfDays }, () => ({
-      activities: Array.from({ length: 3 }, () => getRandomActivity()), // Adjust the number of activities as needed
-    })));
+    setItinerary(
+      Array.from({ length: numberOfDays }, () => ({
+        activities: Array.from({ length: 3 }, () => getRandomActivity()), // Adjust the number of activities as needed
+      }))
+    );
   };
 
   const generateItinerary = (numberOfDays) => {
-    setItinerary(Array.from({ length: numberOfDays }, (_, index) => ({
-      activities: [],
-    })));
+    setItinerary(
+      Array.from({ length: numberOfDays }, (_, index) => ({
+        activities: [],
+      }))
+    );
   };
 
   const handleAddActivity = (day, newActivity) => {
@@ -211,7 +95,10 @@ const ItineraryPage = () => {
   const handleEditActivity = (day, activityIndex) => {
     // Implement your logic for editing an activity here
     // For simplicity, let's just add a placeholder logic
-    const updatedActivity = prompt('Edit activity:', itinerary[day - 1].activities[activityIndex]);
+    const updatedActivity = prompt(
+      "Edit activity:",
+      itinerary[day - 1].activities[activityIndex]
+    );
     if (updatedActivity) {
       setItinerary((prevItinerary) => {
         const updatedItinerary = [...prevItinerary];
@@ -235,7 +122,7 @@ const ItineraryPage = () => {
       <h1>Itinerary</h1>
       <h2>Current City: {currentCity}</h2>
 
-      {currentStep === 'datePicker' && (
+      {currentStep === "datePicker" && (
         <DatePickerComponent
           onSelectStartDate={handleStartDateSelect}
           onSelectEndDate={handleEndDateSelect}
@@ -243,7 +130,7 @@ const ItineraryPage = () => {
         />
       )}
 
-      {currentStep === 'itinerary' && (
+      {currentStep === "itinerary" && (
         <>
           {itinerary.map((day, index) => (
             <ItineraryAccordion
